\name{riojaExtra-package}
\alias{riojaExtra-package}
\alias{riojaExtra}
\docType{package}
\title{
Helper functions for rioja (Analysis of Quaternary Science Data)
}
\description{
\code{riojaExtra}: An Additional functions for the rioja package to import Cornell and Tilia files.
}

\details{
The \emph{riojaExtra} package contains additional function to suppliment the rioja package.

\packageIndices{riojaExtra}
}

\author{
Steve Juggins

Maintainer: Steve Juggins <Stephen.Juggins@ncl.ac.uk>
}
\keyword{ package }

